Pending FInal Checks By Mike, Ready to proceed to Upper Modules 6 and above.
git-tfs-id: [http://tfs00.element.root.com:8080/tfs/TFSCollection]$/VECTO;C1660

+++ VECTOAux/VectoAuxiliaries/Electrics/IAlternatorMap.vb
@@ -15,7 +15,7 @@ Namespace Electrics
         ''' <param name="rpm">alternator rotation speed</param>
         ''' <returns>Single</returns>
         ''' <remarks></remarks>
-        Function GetEfficiency(ByVal rpm As Integer, ByVal amps As Integer) As AlternatorMapValues
+        Function GetEfficiency(ByVal rpm As single, ByVal amps As single) As AlternatorMapValues
 
     End Interface
